+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- Plugin Exception in case of Launch creation timed out, by @HardNorth
+### Changed
+- Client version updated on [5.3.1](https://github.com/reportportal/client-Python/releases/tag/5.3.1), by @HardNorth
+
+## [5.1.6]
+### Changed
+- Client version updated on [5.3.0](https://github.com/reportportal/client-Python/releases/tag/5.3.0), by @HardNorth
 
 ## [5.1.5]
 ### Added